Fix parallel plugin rename.

+++ crates/warpgate/src/helpers.rs
@@ -78,27 +78,36 @@ pub fn move_or_unpack_download(
             return Err(WarpgateLoaderError::NoWasmFound {
                 path: temp_file.to_path_buf(),
             });
-
-            // Find a release file first, as some archives include the target folder
-        } else if let Some(release_wasm) = wasm_files
+        }
+        // Find a release file first, as some archives include the target folder
+        else if let Some(release_wasm) = wasm_files
             .iter()
             .find(|file| file.to_string_lossy().contains("release"))
         {
             fs::rename(release_wasm, dest_file)?;
-
-            // Otherwise, move the first wasm file available
-        } else {
+        }
+        // Otherwise, move the first wasm file available
+        else {
             fs::rename(&wasm_files[0], dest_file)?;
         }
 
         fs::remove_file(temp_file)?;
         fs::remove_dir_all(out_dir)?;
+
+        return Ok(());
     }
 
     // Non-archive supported extensions
     match temp_file.extension().and_then(|ext| ext.to_str()) {
         Some("wasm" | "toml" | "json" | "jsonc" | "yaml" | "yml") => {
-            fs::rename(temp_file, dest_file)?;
+            // Plugins can be downloaded in parallel, which means
+            // that his temp file can also be moved by another process.
+            // Because of this, proto constantly runs into "Failed to rename"
+            // errors when hitting this block, so let's avoid the failure
+            // if the condition is met and assume all is good!
+            if temp_file.exists() && !dest_file.exists() {
+                fs::rename(temp_file, dest_file)?;
+            }
         }
 
         Some(ext) => {
